requirements.txt
youtube-transcript-api
deep-translator
pyyaml
